infra/airflow/pyproject.toml
[project]
name = "airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow==2.11.0",
    "apache-airflow-providers-celery==3.12.0",
    "apache-airflow-providers-docker==4.4.0",
    "apache-airflow-providers-postgres>=6.2.0",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
]
